server/requirements.txt
# Main API server dependencies
Flask>=2.0.0
Pillow>=11.0,<12.0
numpy>=2.0,<2.3
opencv-python>=4.5.3
requests>=2.25.0

# Chess recognition dependencies (from chesscog)
torch>=2.7.1
torchvision>=0.22.1
matplotlib>=3.4.3
tensorboard>=2.6.0
pandas>=2.3.1
scikit-learn>=1.3.2
tqdm>=4.62.2
googledrivedownloader>=0.4
recap>=0.1.6
chess>=1.6.1
osfclient>=0.0.5
setuptools==59.5.0

# local lib chess recogniser (not official)
# chess_positions

# Roboflow detection dependencies
inference-sdk>=0.51.3,<0.52.0

# Optional documentation dependencies (uncomment if needed)
# sphinx>=4.1.2
# sphinx-rtd-theme>=0.5.2
# m2r2>=0.3.1
